Re #423: Split out the batching mechanism from ReplicationSender
-this will eventually be re-used for both ReplicationMessage and ReplicationMessageAck
